Define 'trait QueryEngine' to refactor runtime crate
